travis: Test only latest Ruby and Puppet.
Since this module contains no Ruby code there's really no point in
running a matrix for all Ruby versions. We trust Puppet to work on the
different Ruby versions so all we need is to run the different versions
of Puppet.

We use Ruby 2.1.6 since 2.2.x causes Puppet 3 and 4 to crash.

We need to allow Puppet 4 builds to break right now because rspec-puppet
doesn't work with it.