disable by default and deprecate dpkg "options" parameter
unattended-upgrades does *not* need special configuration to `dpkg`
for proper operation. It handles those itself. All that parameter does
is changes the `dpkg` configuration for *other* `dpkg` runs.

This will, for example, affect major upgrades, typically done by hand,
and not unattended-upgrades, which was *not* designed for that
purpose. In those cases, unsuspecting users will see their old
configuration file versions silently kept in place, which most
definitely *will* break things on upgrade, rather noisily in the end.

People unfamiliar with this mechanism will have a hard time figuring
out what has happened, and while recovery is possible, it is tricky
without special tooling. I wrote such a tool here:

https://gitlab.com/anarcat/koumbit-scripts/-/blob/master/vps/clean_conflicts

... but it not well known and definitely not shipped with Debian by
default.

While it is nice that the unattended-upgrades module allows users to
change dpkg options, this can be more easily (and directly) done with
the `apt::conf` parameter in the first place.

This also marks the feature as deprecated. Once a proper deprecation
delay has passed, the code can be removed (with, for example, PR #190).