Fix security origins on Debian releases past bullseye
Debian 11 (bullseye) changed the codename in the security Release file
to specify <codename>-security instead of just <security>.

For bullseye this was already correctly handled, but for later releases
this module would fall back to the earlier style.

Fix origins to add -security for all releases >= bullseye and not just
for exactly bullseye.